capstone_project/requirements.txt
alembic==1.0.11
apache-airflow==1.10.4
apispec==2.0.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
bleach==3.1.0
boto3==1.7.84
botocore==1.10.84
cached-property==1.5.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
configparser==3.5.3
confuse==1.0.0
croniter==0.3.30
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docutils==0.15.2
dumb-init==1.2.2
entrypoints==0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-AppBuilder==1.13.1
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-JWT-Extended==3.21.0
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gunicorn==19.9.0
htmlmin==0.1.12
idna==2.8
importlib-metadata==0.19
ipykernel==5.1.2
ipython==7.7.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jmespath==0.9.4
json-merge-patch==0.2
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
llvmlite==0.29.0
lockfile==0.12.2
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.19.5
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.17.0
matplotlib==3.1.1
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
numba==0.45.1
numpy==1.17.0
ordereddict==1.1
packaging==19.1
pandas==0.25.0
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.5.1
pendulum==1.4.4
pexpect==4.7.0
phik==0.9.8
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.12.0
prison==0.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psutil==5.6.3
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.0
pytest-pylint==0.14.1
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.2
pytzdata==2019.2
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.3
requests==2.22.0
s3transfer==0.1.13
scipy==1.3.1
seaborn==0.9.0
Send2Trash==1.5.0
setproctitle==1.1.10
six==1.12.0
SQLAlchemy==1.3.7
tabulate==0.8.3
tenacity==4.12.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thrift==0.11.0
tornado==6.0.3
tqdm==4.33.0
traitlets==4.3.2
typed-ast==1.4.0
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.11.2
WTForms==2.2.1
zipp==0.5.2
zope.deprecation==4.4.0
